requirements.txt
streamlit
SpeechRecognition
pyaudio
setuptools
langchain
faiss-cpu
langchain-core
langchain-community
langchain-experimental
langchain-openai
openai
elevenlabs
flask 
twilio
google-cloud-speech
flask_sock
opencv-python
easyocr
google-cloud-vision
pandas